requirements.txt
-r requirements/analyze.txt
-r requirements/dev.txt
-r requirements/test.txt

async-exit-stack==1.0.1
async-generator==1.10
fastapi==0.60.1
motor==2.1.0
passlib==1.7.2
python-jose==3.2.0
python-multipart==0.0.5
pydantic==1.6.1
starlette==0.13.6
uvloop==0.14.0
uvicorn==0.11.7
databases==0.3.2
bcrypt==3.1.7
